requirements.txt
mypy==0.761
pytest==5.0.0
piprot==0.9.11
pytest-cov==2.8.1
flake8==3.7.9
pip-licenses==2.0.1
pydantic==1.3
marshmallow-dataclass==7.2.1
dataclasses-json==0.5.2
dataclasses-avroschema==0.20.2